requirements.txt
yt-dlp
webvtt-py
pydub
aksharamukha
pyspellchecker
unidecode
static-ffmpeg
